reporter: rename parse_result, move badge management to process_result
